client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "gradio>=5.33.0",
    "groq>=0.26.0",
    "mcp>=1.9.3",
    "python-dotenv>=1.1.0",
]
